+++ src/main/java/org/jenkinsci/plugins/codesonar/CodeSonarPublisher.java
@@ -98,6 +98,31 @@ public class CodeSonarPublisher extends Recorder implements SimpleBuildStep  {
         return this.visibilityFilter;
     }
 
+    // projectName may be a
+    private String getAid(@Nonnull Run<?,?> run, @Nonnull FilePath workspace, @Nonnull String projectName) throws java.io.IOException {
+        /**
+         * The analysis ID from the aid.txt file
+         */
+        String aid = workspace.asCallableWith(new hudson.FilePath.FileCallable<String>() {
+            @Override
+            public void checkRoles(org.jenkinsci.remoting.RoleChecker roleChecker) throws SecurityException {
+
+            }
+
+            @Override
+            public String invoke(java.io.File file, hudson.remoting.VirtualChannel virtualChannel) throws java.io.IOException {
+                Logger logger = Logger.getLogger("jenkins.codesonar.plugin");
+                String aid = new String(
+                        java.nio.file.Files.readAllBytes(java.nio.file.Paths.get(file.toString(), projectName + ".prj_files/aid.txt")),
+                        java.nio.charset.StandardCharsets.UTF_8
+                );
+                return aid;
+            }
+        }).call();
+
+        return aid;
+    }
+
     @Override
     public void perform(@Nonnull Run<?, ?> run, @Nonnull FilePath workspace, @Nonnull Launcher launcher,
             @Nonnull TaskListener listener)
@@ -135,14 +160,16 @@ public class CodeSonarPublisher extends Recorder implements SimpleBuildStep  {
         List<String> logFile = IOUtils.readLines(run.getLogReader());
         String analysisUrl = analysisService.getAnalysisUrlFromLogFile(logFile);
 
-        if (analysisUrl == null) {
-            analysisUrl = analysisService.getLatestAnalysisUrlForAProject(baseHubUri, expandedProjectName);
+        String aid = this.getAid(run, workspace, expandedProjectName);
+        if(aid == null){
+            // Can't get the analysis ID, so let's mark the build as aborted
+            run.setResult(hudson.model.Result.ABORTED);
         }
-        // here we also need to change something
+
         Analysis analysisWarnings = analysisService.getAnalysisFromUrlWarningsByFilter(analysisUrl);
-        URI metricsUri = metricsService.getMetricsUriFromAnAnalysisId(baseHubUri, analysisWarnings.getAnalysisId());
+        URI metricsUri = metricsService.getMetricsUriFromAnAnalysisId(baseHubUri, aid);
         Metrics metrics = metricsService.getMetricsFromUri(metricsUri);
-        URI proceduresUri = proceduresService.getProceduresUriFromAnAnalysisId(baseHubUri, analysisWarnings.getAnalysisId());
+        URI proceduresUri = proceduresService.getProceduresUriFromAnAnalysisId(baseHubUri, aid);
         Procedures procedures = proceduresService.getProceduresFromUri(proceduresUri);
 
         Analysis analysisNewWarnings = analysisService.getAnalysisFromUrlWithNewWarnings(analysisUrl);